chore(e2e): move bench deep-link protocol into src/__automation__
The URL prefix, prefix matcher, and autostart query parser were living
in src/utils/navigationConstants.ts so prod-reachable navigation code
co-owned the E2E benchmark protocol. Move them to a new module under
src/__automation__/benchmarkRoute.ts so the automation protocol sits
inside the same boundary as the runner screen and the dispatch helpers.

ROUTES.BENCHMARK_RUNNER stays in the ROUTES enum — it is just a screen
identifier, not protocol surface. The two prod-reachable consumers
(useDeepLinking and App.tsx) are already allow-listed by the ESLint
no-restricted-imports rule, so this move uses that escape hatch for
exactly what it was carved out for.

Tests move along with the symbols; the per-file navigationConstants
test file is now empty and removed.
